fixed opportunity edit page when not logged in
